# optimize processSerializerClass foundTypes